requirements.txt
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
django-filter==25.1
djangorestframework-simplejwt==5.5.1
pandas==2.3.1
Pillow==11.0.0
redis==5.2.1
celery==5.4.0